+++ features/sessions/ui/src/commonMain/kotlin/app/campfire/sessions/ui/ExpandedPlaybackBar.kt
@@ -617,7 +617,7 @@ private fun PlaybackActions(
       )
     }
 
-    val isPlayPauseEnabled = state != AudioPlayer.State.Disabled &&
+    val isPlayPauseEnabled = state != AudioPlayer.State.Finished &&
       state != AudioPlayer.State.Buffering &&
       !isInteracting